requirements.txt

pandas
matplotlib
pyyaml
